Visual mode: added "Look Through Selection" action (default key is "Y"). This action places visual camera at the same position as selected/highlighted thing and rotates it to match thing's angle. Special handling is available if targeted thing is AimingCamera, MovingCamera or SecurityCamera. Script editor: pressing "F1" now opens keyword help instead of program manual. Linedef edit form, Thing edit form: rewritten script support logic. It should now work in the same manner as the rest of controls. Thing info panel: thing arguments now have proper labels. Edit forms, UDMF: fields, which are handled by UI, are no longer shown in "Custom" tab. Visual mode: fixed a crash when loading a model on a video card without Shader model 2.0 support. Fixed incorrect argument number and changed most of labels in "Cameras and interpolation" section of zdoom_things.cfg.

+++ branches/GZDoomBuilder/Source/Core/Rendering/Renderer3D.cs
@@ -305,7 +305,7 @@ namespace CodeImp.DoomBuilder.Rendering
 
 			// Create frustum
 			frustum = new ProjectedFrustum2D(pos, anglexy, anglez, PROJ_NEAR_PLANE,
-				General.Settings.ViewDistance, Angle2D.DegToRad((float)General.Settings.VisualFOV));
+				General.Settings.ViewDistance, Angle2D.DegToRad(General.Settings.VisualFOV));
 			
 			// Make the view matrix
 			view3d = Matrix.LookAtRH(D3DDevice.V3(pos), D3DDevice.V3(lookat), new Vector3(0f, 0f, 1f));
